requirements.txt
pyyaml
mediapipe
opencv-python
streamlit
matplotlib
numpy